Updated application to dotnetcore 2.0
