+++ Knez/Knez.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.0.0
+MinimumVisualStudioVersion = 10.0.0.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Knez", "Knez\Knez.csproj", "{09C59998-6486-4029-BE18-2A6BD379A8A2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{09C59998-6486-4029-BE18-2A6BD379A8A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{09C59998-6486-4029-BE18-2A6BD379A8A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{09C59998-6486-4029-BE18-2A6BD379A8A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{09C59998-6486-4029-BE18-2A6BD379A8A2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		version = 2.0.0
+	EndGlobalSection
+EndGlobal
