+++ Knez/Knez/Knez.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <ReleaseVersion>2.0.0</ReleaseVersion>
+  </PropertyGroup>
+
+</Project>
